# Fix array performance regression