* `ObjectChangeFlags.NonDefaultDataDescriptorUsage `should only be raised for array index
* inline some unnecessary extra methods
* reuse property descriptor if possible on `Set`