@lahma 2022-09-04
## Jint.Benchmark.ArrayStressBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |Execute|13.40 ms|0.085 ms|8.34 MB|
| **New** |	| **11.18 ms (-17%)** | **0.128 ms** | **8.14 MB (-2%)** |
| Old |Execute_ParsedScript|13.47 ms|0.037 ms|8.32 MB|
| **New** |	| **10.87 ms (-19%)** | **0.056 ms** | **8.12 MB (-2%)** |